Do not close notifier with mvnd
https://github.com/mvndaemon/mvnd is a Gradle like daemon for Maven.

When used with maven-notifer, a notification is only displayed during the first build
on a fresh daemon.
For succeeding builds, the notification will not be displayed.

It comes from the fact that SendNotificationNotifier will never be re-instantiated after the 'close'
method has been called.
In a standard Maven lifecycle everything is startup for each build as a new JVM is stared.
But this is not the case with mvnd.

Notifiers should not rely on state given by the constructor (meaning re-initializing everything in 'init') or do not 'close' at all and let resources being cleaned up when the daemon stop.

This commit does not call 'close' anymore when we detect that a build is run with mvnd.

Fixes #108